+++ django_mapengine/utils.py
@@ -1,28 +1,56 @@
-from itertools import product
+"""Various function to be used in other modules"""
+from typing import Union
 
 from django.conf import settings
-from django.db.models import BooleanField, IntegerField
 
 
-def get_color(source_layer):
-    if source_layer not in settings.MAP_ENGINE_LAYER_STYLES:
-        raise KeyError(f"Could not find layer '{source_layer}' in layer styles (static/config/layer_styles.json)")
+def get_color(layer_id: str) -> Union[str, dict]:
+    """
+    Return color of given layer
+
+    Parameters
+    ----------
+    layer_id : str
+        Layer ID to lookup up in layer styles
+
+    Returns
+    -------
+    Union[str, dict]
+        Color of given layer
+
+    Raises
+    ------
+    LookupError
+        if color is not found in layer style
+    """
+    style = get_layer_style(layer_id)
     for color_key in ("fill-color", "line-color", "circle-color"):
         try:
-            return settings.MAP_ENGINE_LAYER_STYLES[source_layer]["paint"][color_key]
+            return style["paint"][color_key]
         except KeyError:
             continue
-    return None
+    raise LookupError(f"Could not find color for {layer_id=}.")
 
 
-def get_layer_setups(layer):
-    setups = []
-    setup_model = layer.model._meta.get_field("setup").related_model
-    for setup in setup_model._meta.fields:
-        if setup.name == "id":
-            continue
-        if isinstance(setup, IntegerField):
-            setups.append([f"{setup.name}={choice[0]}" for choice in setup.choices])
-        elif isinstance(setup, BooleanField):
-            setups.append([f"{setup.name}=True", f"{setup.name}=False"])
-    return product(*setups)
+def get_layer_style(layer_id: str) -> dict:
+    """
+    Return layer style of given layer ID
+
+    Parameters
+    ----------
+    layer_id : str
+        Layer ID to lookup style
+
+    Returns
+    -------
+    dict
+        Layer style of given ID
+
+    Raises
+    ------
+    LookupError
+        if layer ID is not found in styles file
+    """
+    if layer_id not in settings.MAP_ENGINE_LAYER_STYLES:
+        raise LookupError(f"Could not find '{layer_id=}' in layer styles.")
+    return settings.MAP_ENGINE_LAYER_STYLES[layer_id]